Learn the mysteries of RSpec from Tomas
