+++ spec/lib/dfe/reference_data/degrees/combined_subjects_spec.rb
@@ -1,3 +1,5 @@
+require 'support/autocomplete'
+
 RSpec.describe DfE::ReferenceData::Degrees::COMBINED_SUBJECTS do
   describe 'subject IDs are correct' do
     let(:records) { described_class.all }
@@ -36,9 +38,5 @@ RSpec.describe DfE::ReferenceData::Degrees::COMBINED_SUBJECTS do
     end
   end
 
-  describe 'combined subjects' do
-    it 'is a valid autocomplete-capable list' do
-      DfE::ReferenceData::Degrees::COMBINED_SUBJECTS.validate_autocomplete_compatibility!
-    end
-  end
+  it_should_behave_like 'a valid autocomplete-capable list'
 end
